requirements.txt
flask
python-dotenv
line-bot-sdk>=3.0.0,<4.0.0
qdrant-client
langchain
langchain-openai
langchain-qdrant
langchain-community
openai
tiktoken
pypdf